Move OutputTableItem to separate OutputsUtils.py in preparation for refactoring AddOns/S-DAC output/rich output flow
