+++ src/geophires_x/OutputsS_DAC_GT.py
@@ -1,7 +1,7 @@
 import sys
 import pandas as pd
 from geophires_x.Outputs import Outputs
-from geophires_x.OutputsRich import OutputTableItem
+from geophires_x.OutputsUtils import OutputTableItem
 
 NL = "\n"
